aiagent/requirements1.txt
python-dotenv
livekit-agents
livekit
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-cartesia
livekit-plugins-silero
livekit-plugins-turn-detector
